+++ src/browser/modules/DBMSInfo/DatabaseSelector.test.tsx
@@ -21,7 +21,7 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
 import { DatabaseSelector } from './DatabaseSelector'
-import { Database } from 'shared/modules/dbMeta/types'
+import { Database } from 'shared/modules/dbMeta/state'
 
 const testId = 'database-selection-list'
