+++ src/browser/modules/Stream/SysInfoFrame/SysInfoFrame.tsx
@@ -22,8 +22,11 @@ import React, { Component, ReactNode } from 'react'
 import { connect } from 'react-redux'
 import { withBus } from 'react-suber'
 import { CYPHER_REQUEST } from 'shared/modules/cypher/cypherDuck'
-import { isEnterprise } from 'shared/modules/dbMeta/selectors'
-import { Database } from 'shared/modules/dbMeta/types'
+import {
+  isEnterprise,
+  Database,
+  getDatabases
+} from 'shared/modules/dbMeta/state'
 import {
   isConnected,
   getUseDb
@@ -38,7 +41,6 @@ import {
 import { NEO4J_BROWSER_USER_ACTION_QUERY } from 'services/bolt/txMetadata'
 import { hasMultiDbSupport } from 'shared/modules/features/versionedFeatures'
 import { ErrorsView } from '../CypherFrame/ErrorsView'
-import { getDatabases } from 'shared/modules/dbMeta/selectors'
 import * as legacyHelpers from './legacyHelpers'
 import * as helpers from './helpers'
 import { SysInfoTable } from './SysInfoTable'
