+++ src/browser/modules/User/UserList.tsx
@@ -42,7 +42,7 @@ import { forceFetch } from 'shared/modules/currentUser/currentUserDuck'
 import { NEO4J_BROWSER_USER_ACTION_QUERY } from 'services/bolt/txMetadata'
 import { driverDatabaseSelection } from 'shared/modules/features/versionedFeatures'
 import { connect } from 'react-redux'
-import { isEnterprise } from 'shared/modules/dbMeta/selectors'
+import { isEnterprise } from 'shared/modules/dbMeta/state'
 import FrameAside from '../Frame/FrameAside'
 import { EnterpriseOnlyFrame } from 'browser-components/EditionView'
 import { isConnectedAuraHost } from 'shared/modules/connections/connectionsDuck'
