+++ src/shared/modules/commands/helpers/config.test.ts
@@ -21,7 +21,7 @@
 import nock from 'nock'
 import * as config from './config'
 import { update, replace } from 'shared/modules/settings/settingsDuck'
-import dbMetaReducer from 'shared/modules/dbMeta/reducer'
+import dbMetaReducer from 'shared/modules/dbMeta/state'
 import { updateSettings } from 'shared/modules/dbMeta/actions'
 
 function FetchError(message: any) {
