+++ src/shared/modules/commands/helpers/server.ts
@@ -22,7 +22,7 @@ import { getCommandAndParam } from 'services/commandUtils'
 import * as connections from 'shared/modules/connections/connectionsDuck'
 import { add as addFrameAction } from 'shared/modules/frames/framesDuck'
 import { UnknownCommandError } from 'services/exceptions'
-import { shouldRetainConnectionCredentials } from 'shared/modules/dbMeta/selectors'
+import { shouldRetainConnectionCredentials } from 'shared/modules/dbMeta/state'
 import { clearRefreshTokenData } from 'neo4j-client-sso'
 
 export function handleServerCommand(action: any, put: any, store: any) {
