--- src/shared/modules/dbMeta/types.ts
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-export type Database = {
-  name: string
-  address: string
-  role: string
-  requestedStatus: string
-  currentStatus: string
-  error: string
-  default: boolean
-  home?: boolean // introduced in neo4j 4.3
-  aliases?: string[] // introduced in neo4j 4.4
-  status: string
-}
