+++ src/shared/modules/features/featuresDuck.test.ts
@@ -20,7 +20,7 @@
 
 import reducer, * as features from './featuresDuck'
 import { canSendTxMetadata } from './versionedFeatures'
-import { NAME as META_NAME } from '../dbMeta/constants'
+import { NAME as META_NAME } from '../dbMeta/state'
 import { dehydrate } from 'services/duckUtils'
 
 describe('features reducer', () => {
